Login/logout issue fix  (#274)
* build error fix

* build error fix

* build error fix

* build error fix

* review chnages

* review chnages
